+++ src/Zones/ZoneFeaturePublisher.cs
@@ -8,7 +8,7 @@ namespace Zones;
 
 public interface IFeatureBus 
 { 
-    void Publish(string symbol, DateTime utc, string name, double value); 
+    void Publish(string symbol, DateTime utc, string name, decimal value); 
 }
 
 public sealed class ZoneFeaturePublisher : BackgroundService
